remove single-use deduper mixin
